+++ src/jk_utils/datetime/__init__.py
@@ -1,7 +1,7 @@
 ﻿
 
 
-__version__ = "0.2021.1.11.1"
+__version__ = "0.2021.1.13"
 
 
 from .D import D