+++ src/jk_utils/tokenizer2/__init__.py
@@ -1,6 +1,6 @@
 ﻿
 
-__version__ = "0.2021.1.11.1"
+__version__ = "0.2021.1.13"
 
 
 from .Token import Token